requirements.txt
pynwb ~= 2.3
nwbinspector ~= 0.4
neo ~= 0.12
numpy
pendulum
pandas
pyyaml
opencv-python-headless
Pillow
imageio
av
fastcrc
#open-ephys-python-tools https://github.com/open-ephys/open-ephys-python-tools